requirements.txt
pandas
openpyxl
python-dotenv
fpdf
qrcode
pillow
gunicorn

# pip install -r requirements.txt